+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The following changes have been implemented but not released yet:
 
 - The function `getLinkedResourceUrlAll` that gives you all Resources linked to
   a given Resource, indexed by their relation to the given Resource.
+- The function `getEffectiveAccess`, which tells you what access the current
+  user has and, if supported by the server, what access unauthenticated users
+  have to the given Resource.
 
 ### Bugs fixed